langchain-fastapi-demo/requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
langchain>=0.2.12
langchain-openai>=0.1.6
python-dotenv>=1.0.1